edgecore/dht/app/requirements.txt
Adafruit-DHT --install-option="--force-pi2"
paho-mqtt
argparse